backend/requirements.txt
deepgram-sdk==5.0.0
docx2txt==0.9
elevenlabs==2.18.0
Flask[async]==3.1.2
flask-cors==6.0.1
google-cloud-texttospeech==2.31.0
google-generativeai==0.8.5
groq==0.32.0
langchain==0.3.27
langchain-cli==0.0.37
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-openai==0.3.35
langchain-qdrant==0.2.1
langchain-unstructured==0.1.6
langdetect==1.0.9
langgraph==0.6.10
langserve==0.3.2
langsmith==0.4.36
mysql-connector-python==9.4.0
openai-whisper==20250625
openpyxl==3.1.5
pandas==2.3.3
Pillow==12.0.0
PyMuPDF==1.26.5
PyMySQL==1.1.2
pytesseract==0.3.13
python-dotenv==1.1.1
qdrant-client==1.15.1
sarvamai==0.1.21
sentence-transformers==5.1.1
soxr==1.0.0
SQLAlchemy==2.0.44
torch==2.9.0
torchaudio==2.9.0
unstructured[pptx,pdf,docx]==0.18.15
spacy
mcp
duckduckgo-search